Update Arena.php
Literally the first line(s) of PHP i've ever written so I have no idea if it is correct or not, but it is intended to set players gamemode to survival before they enter the arena. I thought that would be pretty nice because on my server that was the only problem.